use correct type

+++ src/serac/physics/utilities/finite_element_state.cpp
@@ -70,7 +70,7 @@ double avg(const FiniteElementState& state)
   double global_sum;
   double local_sum = state.trueVec().Sum();
   int    global_size;
-  double local_size = state.trueVec().Size();
+  int    local_size = state.trueVec().Size();
   MPI_Allreduce(&local_sum, &global_sum, 1, MPI_DOUBLE, MPI_SUM, state.comm());
   MPI_Allreduce(&local_size, &global_size, 1, MPI_INT, MPI_SUM, state.comm());
   return global_sum / global_size;
